working on baggers
