+++ src/main/scala/io/citrine/lolo/bags/BaggedModel.scala
@@ -1,0 +1,5 @@
+package io.citrine.lolo.bags
+
+import io.citrine.lolo.Model
+
+trait BaggedModel[T] extends Model[T] {}
