+++ src/main/scala/io/citrine/lolo/bags/BaggedResult.scala
@@ -1,20 +1,15 @@
 package io.citrine.lolo.bags
 
-import breeze.linalg.{norm, DenseMatrix, DenseVector}
-import io.citrine.lolo.stats.StatsUtils
-import io.citrine.lolo.{
-  MultiTaskModelPredictionResult,
-  ParallelModelsPredictionResult,
-  PredictionResult,
-  RegressionResult
-}
+import breeze.linalg.DenseMatrix
+import io.citrine.lolo.PredictionResult
 import org.slf4j.{Logger, LoggerFactory}
 
 /**
   * Interface defining the return value of a [[BaggedModel]].
   *
-  * This allows the implementation to depend on the number of simultaneous predictions, which has performance
-  * implications.
+  * This allows the implementation to depend on the number of simultaneous predictions,
+  * which has performance implications.
+  *
   * For background on the uncertainty calculation, see Wager, S.; Hastie, T and Efron, B. Confidence Intervals for
   * Random Forests: The Jackknife and Infinitesimal Jackknife. Journal of Machine Learning Research 15 (2014).
   */
@@ -43,434 +38,6 @@ trait BaggedResult[+T] extends PredictionResult[T] {
       r.toVector.transpose.map(_.sum / predictions.size)
     })
   }
-}
-
-/**
-  * Container with model-wise predictions at a single input point.
-  * Assuming a single input allows for performance optimizations and more readable code.
-  * See [[MultiPredictionBaggedResult]] for a generic implementation.
-  *
-  * @param predictions for each constituent model
-  * @param NibIn       the sample matrix as (N_models x N_training)
-  * @param bias        model to use for estimating bias
-  */
-case class SinglePredictionBaggedResult(
-    predictions: Seq[PredictionResult[Double]],
-    NibIn: Vector[Vector[Int]],
-    bias: Option[Double] = None,
-    rescale: Double = 1.0,
-    disableBootstrap: Boolean = false
-) extends BaggedResult[Double]
-    with RegressionResult {
-  private lazy val treePredictions: Array[Double] = predictions.map(_.getExpected().head).toArray
-
-  override def numPredictions: Int = 1
-
-  /**
-    * Return the ensemble average or maximum vote
-    *
-    * @return expected value of each prediction
-    */
-  override def getExpected(): Seq[Double] = Seq(expected + bias.getOrElse(0.0))
-
-  private lazy val expected = treePredictions.sum / treePredictions.length
-  private lazy val treeVariance: Double = {
-    assert(treePredictions.length > 1, "Bootstrap variance undefined for fewer than 2 bootstrap samples.")
-    treePredictions.map(x => Math.pow(x - expected, 2.0)).sum / (treePredictions.length - 1)
-  }
-
-  override def getStdDevMean(): Option[Seq[Double]] = {
-    if (disableBootstrap) {
-      // If bootstrap is disabled, rescale is unity and treeVariance is our only option for UQ.
-      // Since it's not recalibrated, it's best considered to be a confidence interval of the underlying weak learner.
-      assert(rescale == 1.0)
-      Some(Seq(Math.sqrt(treeVariance)))
-    } else {
-      Some(Seq(stdDevMean))
-    }
-  }
-
-  override def getStdDevObs(): Option[Seq[Double]] = {
-    if (disableBootstrap) {
-      None
-    } else {
-      Some(Seq(stdDevObs))
-    }
-  }
-
-  /**
-    * For the sake of parity, we were using this method
-    */
-  override def getUncertainty(observational: Boolean): Option[Seq[Any]] = {
-    if (observational) {
-      getStdDevObs()
-    } else {
-      getStdDevMean()
-    }
-  }
-
-  private lazy val stdDevMean: Double = Math.sqrt(BaggedResult.rectifyEstimatedVariance(singleScores))
-
-  private lazy val stdDevObs: Double = {
-    rescale * Math.sqrt(treeVariance)
-  } ensuring (_ >= 0.0)
-
-  /**
-    * The importances are computed as an average of bias-corrected jackknife-after-bootstrap
-    * and infinitesimal jackknife methods
-    *
-    * @return training row scores of each prediction
-    */
-  override def getImportanceScores(): Option[Seq[Seq[Double]]] = Some(Seq(singleScores))
-
-  private lazy val singleScores: Vector[Double] = {
-    // This will be more convenient later
-    val nMat = NibIn.transpose
-
-    // Compute the Bessel-uncorrected variance of the ensemble of predicted values, then multiply by (n - 1) / (n * B)
-    // We later sum over the training data, introducing a factor of n and leaving us with the expected correction term
-    val correction =
-      treeVariance * (treePredictions.length - 1.0) * (nMat.size - 1) / (treePredictions.length * treePredictions.length * nMat.size)
-
-    // Loop over each of the training instances, computing its contribution to the uncertainty
-    val trainingContributions = nMat.indices.toVector.map { idx =>
-      // Pull the vector of the number of times this instance was used to train each tree
-      val vecN = nMat(idx).toArray
-
-      // Loop over the trees, computing the covariance for the IJ estimate and the predicted value of
-      // the out-of-bag trees for the J(ackknife) estimate
-      // The loops are merged for performance reasons
-      var cov: Double = 0.0
-      var tNot: Double = 0.0
-      var tNotCount: Int = 0
-      vecN.indices.foreach { jdx =>
-        cov = cov + (vecN(jdx) - 1) * (treePredictions(jdx) - expected)
-
-        if (vecN(jdx) == 0) {
-          tNot = tNot + treePredictions(jdx)
-          tNotCount = tNotCount + 1
-        }
-      }
-      // Compute the infinitesimal jackknife estimate
-      val varIJ = Math.pow(cov / vecN.length, 2.0)
-
-      if (tNotCount > 0) {
-        // Compute the Jackknife after bootstrap estimate
-        val varJ = Math.pow(tNot / tNotCount - expected, 2.0) * (nMat.size - 1) / nMat.size
-        // Averaged the corrected IJ and J estimates
-        0.5 * (varJ + varIJ - Math.E * correction)
-      } else {
-        // We can't compute the Jackknife after bootstrap estimate, so just correct the IJ estimate
-        varIJ - correction
-      }
-    }
-    // The uncertainty must be positive, so anything smaller than zero is noise.  Make sure that no estimated
-    // uncertainty is below that noise level
-    trainingContributions
-  }
-}
-
-case class BaggedClassificationResult[T](predictions: Seq[PredictionResult[T]]) extends BaggedResult[T] {
-  lazy val expectedMatrix: Seq[Seq[T]] = predictions.map(p => p.getExpected()).transpose
-  lazy val expected: Seq[T] = expectedMatrix.map(ps => ps.groupBy(identity).maxBy(_._2.size)._1)
-  lazy val uncertainty: Seq[Map[T, Double]] =
-    expectedMatrix.map(ps => ps.groupBy(identity).view.mapValues(_.size.toDouble / ps.size).toMap)
-
-  override def numPredictions: Int = expectedMatrix.length
-
-  /**
-    * Return the majority vote vote
-    *
-    * @return expected value of each prediction
-    */
-  override def getExpected(): Seq[T] = expected
-
-  override def getUncertainty(includeNoise: Boolean = true): Option[Seq[Map[T, Double]]] = Some(uncertainty)
-}
-
-/**
-  * Container with model-wise predictions and logic to compute variances and training row scores
-  *
-  * These calculations are implemented using matrix arithmetic to make them more performant when the number
-  * of predictions is large.  This obfuscates the algorithm significantly, however.  To see what is being computed,
-  * look at [[SinglePredictionBaggedResult]], which is more clear.  These two implementations are tested for consistency.
-  *
-  * @param predictions for each constituent model
-  * @param NibIn       the sample matrix as (N_models x N_training)
-  * @param bias        model to use for estimating bias
-  */
-case class MultiPredictionBaggedResult(
-    predictions: Seq[PredictionResult[Double]],
-    NibIn: Vector[Vector[Int]],
-    bias: Option[Seq[Double]] = None,
-    rescale: Double = 1.0,
-    disableBootstrap: Boolean = false
-) extends BaggedResult[Double]
-    with RegressionResult {
-
-  /**
-    * Return the ensemble average
-    *
-    * @return expected value of each prediction
-    */
-  override def getExpected(): Seq[Double] = expected.zip(biasCorrection).map(x => x._1 + x._2)
-
-  override def getStdDevObs(): Option[Seq[Double]] = {
-    if (disableBootstrap) {
-      None
-    } else {
-      Some(varObs.map { v => Math.sqrt(v) })
-    }
-  }
-
-  override def getStdDevMean(): Option[Seq[Double]] = {
-    if (disableBootstrap) {
-      // If bootstrap is disabled, rescale is unity and treeVariance is our only option for UQ.
-      // Since it's not recalibrated, it's best considered to be a confidence interval of the underlying weak learner.
-      assert(rescale == 1.0)
-      Some(varObs.map { v => Math.sqrt(v) })
-    } else {
-      Some(stdDevMean)
-    }
-  }
-
-  /**
-    * For the sake of parity, we were using this method
-    */
-  override def getUncertainty(observational: Boolean): Option[Seq[Any]] = {
-    if (observational) {
-      getStdDevObs()
-    } else {
-      getStdDevMean()
-    }
-  }
-
-  /**
-    * Return IJ scores
-    *
-    * @return training row scores of each prediction
-    */
-  override def getInfluenceScores(actuals: Seq[Any]): Option[Seq[Seq[Double]]] = {
-    Some(
-      influences(
-        expected.toVector,
-        actuals.toVector.asInstanceOf[Vector[Double]],
-        expectedMatrix,
-        NibJMat,
-        NibIJMat
-      )
-    )
-  }
-
-  override def getImportanceScores(): Option[Seq[Seq[Double]]] = Some(scores)
-
-  override def numPredictions: Int = expectedMatrix.length
-
-  /* transpose to be (# training) x (# models) */
-  lazy val Nib: Vector[Vector[Int]] = NibIn.transpose
-
-  /* Make a matrix of the tree-wise predictions */
-  lazy val expectedMatrix: Seq[Seq[Double]] = predictions.map(p => p.getExpected()).transpose
-
-  /* Extract the prediction by averaging over trees and adding the bias correction. */
-  lazy val biasCorrection: Seq[Double] = bias.getOrElse(Seq.fill(expectedMatrix.length)(0))
-  lazy val expected: Seq[Double] = expectedMatrix.map(ps => ps.sum / ps.size)
-
-  lazy val NibJMat: DenseMatrix[Double] = BaggedResult.getJackknifeAfterBootstrapMatrix(Nib)
-
-  lazy val NibIJMat: DenseMatrix[Double] = BaggedResult.getInfinitesimalJackknifeMatrix(Nib)
-
-  /* This represents the variance of the estimate of the mean. */
-  lazy val stdDevMean: Seq[Double] =
-    variance(expected.toVector, expectedMatrix, NibJMat, NibIJMat).map { Math.sqrt }
-
-  /* This estimates the variance of predictive distribution. */
-  lazy val varObs: Seq[Double] = expectedMatrix.zip(expected).map {
-    case (b, y) =>
-      assert(Nib.size > 1, "Bootstrap variance undefined for fewer than 2 bootstrap samples.")
-      b.map { x => rescale * rescale * Math.pow(x - y, 2.0) }.sum / (b.size - 1)
-  }
-
-  /* Compute the scores one prediction at a time */
-  lazy val scores: Seq[Vector[Double]] = scores(expected.toVector, expectedMatrix, NibJMat, NibIJMat)
-    // make sure the variance is non-negative after the stochastic correction
-    .map(BaggedResult.rectifyImportanceScores)
-    .map(_.map(Math.sqrt))
-
-  /**
-    * Compute the variance of a prediction as the average of bias corrected IJ and J variance estimates
-    *
-    * @param meanPrediction   over the models
-    * @param modelPredictions prediction of each model
-    * @param NibJ             sampling matrix for the jackknife-after-bootstrap estimate
-    * @param NibIJ            sampling matrix for the infinitesimal jackknife estimate
-    * @return the estimated variance
-    */
-  def variance(
-      meanPrediction: Vector[Double],
-      modelPredictions: Seq[Seq[Double]],
-      NibJ: DenseMatrix[Double],
-      NibIJ: DenseMatrix[Double]
-  ): Seq[Double] = {
-    scores(meanPrediction, modelPredictions, NibJ, NibIJ).map { BaggedResult.rectifyEstimatedVariance }
-  }
-
-  /**
-    * Compute the IJ training row scores for a prediction
-    *
-    * @param meanPrediction   over the models
-    * @param modelPredictions prediction of each model
-    * @param NibJ             sampling matrix for the jackknife-after-bootstrap estimate
-    * @param NibIJ            sampling matrix for the infinitesimal jackknife estimate
-    * @return the score of each training row as a vector of doubles
-    */
-  def scores(
-      meanPrediction: Vector[Double],
-      modelPredictions: Seq[Seq[Double]],
-      NibJ: DenseMatrix[Double],
-      NibIJ: DenseMatrix[Double]
-  ): Seq[Vector[Double]] = {
-    /* Stick the predictions in a breeze matrix */
-    val predMat =
-      new DenseMatrix[Double](modelPredictions.head.size, modelPredictions.size, modelPredictions.flatten.toArray)
-
-    /* These operations are pulled out of the loop and extra-verbose for performance */
-    val JMat = NibJ.t * predMat
-    val JMat2 = JMat *:* JMat * ((Nib.size - 1.0) / Nib.size)
-    val IJMat = NibIJ.t * predMat
-    val IJMat2 = IJMat *:* IJMat
-    val arg = IJMat2 + JMat2
-
-    /* Avoid division in the loop, calculate (n - 1) / (n * B^2) */
-    val prefactor = 1.0 / Math.pow(modelPredictions.head.size, 2.0) * (Nib.size - 1) / Nib.size
-
-    modelPredictions.indices
-      .map { i =>
-        /* Compute the first order bias correction for the variance estimators */
-        val correction = prefactor * Math.pow(norm(predMat(::, i) - meanPrediction(i)), 2)
-
-        /* The correction is prediction dependent, so we need to operate on vectors */
-        0.5 * (arg(::, i) - Math.E * correction)
-      }
-      .map(_.toScalaVector)
-  }
-
-  /**
-    * Compute the IJ training row scores for a prediction
-    *
-    * @param meanPrediction   over the models
-    * @param modelPredictions prediction of each model
-    * @param NibJ             sampling matrix for the jackknife-after-bootstrap estimate
-    * @param NibIJ            sampling matrix for the infinitesimal jackknife estimate
-    * @return the score of each training row as a vector of doubles
-    */
-  def influences(
-      meanPrediction: Vector[Double],
-      actualPrediction: Vector[Double],
-      modelPredictions: Seq[Seq[Double]],
-      NibJ: DenseMatrix[Double],
-      NibIJ: DenseMatrix[Double]
-  ): Seq[Vector[Double]] = {
-    /* Stick the predictions in a breeze matrix */
-    val predMat =
-      new DenseMatrix[Double](modelPredictions.head.size, modelPredictions.size, modelPredictions.flatten.toArray)
-
-    /* These operations are pulled out of the loop and extra-verbose for performance */
-    val JMat = NibJ.t * predMat
-    val IJMat = NibIJ.t * predMat
-    val arg = IJMat + JMat
-
-    /* Avoid division in the loop */
-    val inverseSize = 1.0 / modelPredictions.head.size
-
-    modelPredictions.indices
-      .map { i =>
-        /* Compute the first order bias correction for the variance estimators */
-        val correction = inverseSize * norm(predMat(::, i) - meanPrediction(i))
-
-        /* The correction is prediction dependent, so we need to operate on vectors */
-        val influencePerRow: DenseVector[Double] =
-          Math.signum(actualPrediction(i) - meanPrediction(i)) * 0.5 * (arg(::, i) - Math.E * correction)
-
-        /* Impose a floor in case any of the variances are negative (hacked to work in breeze) */
-        // val floor: Double = Math.min(0, -min(variancePerRow))
-        // val rezero: DenseVector[Double] = variancePerRow - floor
-        // 0.5 * (rezero + abs(rezero)) + floor
-        influencePerRow
-      }
-      .map(_.toScalaVector)
-  }
-}
-
-/**
-  * Container with model-wise predictions for each label and the machinery to compute (co)variance.
-  *
-  * @param baggedPredictions  bagged prediction results for each label
-  * @param realLabels         a boolean sequence indicating which labels are real-valued
-  * @param NibIn              the sampling matrix as (# bags) x (# training)
-  */
-case class MultiTaskBaggedResult(
-    baggedPredictions: Vector[BaggedResult[Any]],
-    realLabels: Seq[Boolean],
-    NibIn: Vector[Vector[Int]]
-) extends BaggedResult[Seq[Any]]
-    with MultiTaskModelPredictionResult {
-
-  /* transpose to be (# training) x (# models) */
-  lazy val Nib: Vector[Vector[Int]] = NibIn.transpose
-
-  override def numPredictions: Int = baggedPredictions.head.numPredictions
-
-  override def getExpected(): Seq[Vector[Any]] = baggedPredictions.map(_.getExpected()).transpose
-
-  override def predictions: Seq[PredictionResult[Seq[Any]]] =
-    baggedPredictions
-      .map(_.predictions.map(_.getExpected()))
-      .transpose
-      .map(x => new ParallelModelsPredictionResult(x.transpose))
-
-  // For each prediction, the uncertainty is a sequence of entries for each label. Missing uncertainty values are reported as NaN
-  override def getUncertainty(observational: Boolean = true): Option[Seq[Seq[Any]]] = {
-    Some(baggedPredictions.map { predictionResult =>
-      predictionResult.getUncertainty(observational) match {
-        case Some(value) => value
-        case None        => Seq.fill(numPredictions)(Double.NaN)
-      }
-    }.transpose)
-  }
-
-  override def getUncertaintyCorrelation(i: Int, j: Int, observational: Boolean = true): Option[Seq[Double]] = {
-    (realLabels(i), realLabels(j)) match {
-      case (true, true) if i == j => Some(Seq.fill(numPredictions)(1.0))
-      case (true, true) =>
-        if (observational) {
-          Some(uncertaintyCorrelationObservational(i, j))
-        } else {
-          Some(uncertaintyCorrelationMean)
-        }
-      case _: Any => None
-    }
-  }
-
-  /** The uncertainty correlation of the observational distribution is the correlation coefficient calculated over the bootstrap ensemble predictions. */
-  private def uncertaintyCorrelationObservational(i: Int, j: Int): Seq[Double] = {
-    // make (# predictions) x (# bags) prediction matrices for each label
-    val baggedPredictionsI =
-      baggedPredictions(i).predictions.map(_.getExpected()).transpose.asInstanceOf[Seq[Seq[Double]]]
-    val baggedPredictionsJ =
-      baggedPredictions(j).predictions.map(_.getExpected()).transpose.asInstanceOf[Seq[Seq[Double]]]
-    baggedPredictionsI.zip(baggedPredictionsJ).map {
-      case (bagsI, bagsJ) =>
-        StatsUtils.correlation(bagsI, bagsJ)
-    }
-  }
-
-  /**
-    * The uncertainty correlation of the mean distribution is 0.0. In theory it should be estimated using the jackknife,
-    * but in practice the jackknife performs poorly when estimating covariance, so we default to the trivial implementation for now.
-    */
-  private def uncertaintyCorrelationMean: Seq[Double] = Seq.fill(numPredictions)(0.0)
-
 }
 
 object BaggedResult {
