+++ src/main/scala/io/citrine/lolo/bags/BaggedTrainingResult.scala
@@ -1,0 +1,5 @@
+package io.citrine.lolo.bags
+
+import io.citrine.lolo.TrainingResult
+
+trait BaggedTrainingResult[+T] extends TrainingResult[T] {}
