+++ src/main/scala/io/citrine/lolo/bags/BaggerHelper.scala
@@ -15,35 +15,26 @@ import scala.collection.parallel.immutable.ParSeq
   * @param uncertaintyCalibration whether to apply empirical uncertainty calibration
   */
 protected case class BaggerHelper(
-    models: ParSeq[Model[Any]],
-    trainingData: Seq[(Vector[Any], Any)],
+    models: ParSeq[Model[Double]],
+    trainingData: Seq[(Vector[Any], Double)],
     Nib: Vector[Vector[Int]],
     useJackknife: Boolean,
     uncertaintyCalibration: Boolean
 ) {
-  val isRegression: Boolean = trainingData.map { _._2 }.find { _ != null } match {
-    case Some(_: Double) => true
-    case Some(_: Any)    => false
-    case None            => throw new IllegalArgumentException(s"Unable to find a non-null label")
-  }
 
   /**
     * Seq of tuples containing training data paired with their out-of-bag residuals and uncertainty estimates.
     * Defined as lazy so it's only computed if needed for the ratio or bias learner calculations.
     */
   lazy val oobErrors: Seq[(Vector[Any], Double, Double)] = trainingData.indices.flatMap { idx =>
-    val oobModels = models.zip(Nib.map(_(idx))).filter(_._2 == 0).map(_._1).asInstanceOf[ParSeq[Model[Double]]]
+    val oobModels = models.zip(Nib.map(_(idx))).filter(_._2 == 0).map(_._1)
     val label = trainingData(idx)._2
-    if (oobModels.size < 2 || label == null || (label.isInstanceOf[Double] && label.asInstanceOf[Double].isNaN)) {
+    if (oobModels.size < 2 || label.isNaN) {
       None
     } else {
-      val model = new BaggedModel(
-        oobModels,
-        Nib.filter { _(idx) == 0 },
-        useJackknife
-      )
+      val model = new BaggedRegressionModel(oobModels, Nib.filter { _(idx) == 0 })
       val predicted = model.transform(Seq(trainingData(idx)._1))
-      val error = predicted.getExpected().head - trainingData(idx)._2.asInstanceOf[Double]
+      val error = predicted.getExpected().head - trainingData(idx)._2
       val uncertainty = predicted match {
         case x: RegressionResult => x.getStdDevObs().get.head
         case _: Any =>
@@ -54,8 +45,8 @@ protected case class BaggerHelper(
   }
 
   /** Uncertainty calibration ratio based on the OOB errors. */
-  val rescaleRatio: Double = if (uncertaintyCalibration && isRegression && useJackknife) {
-    val trainingLabels = trainingData.collect { case (_, x: Double) if !(x.isInfinite || x.isNaN) => x }
+  val rescaleRatio: Double = if (uncertaintyCalibration && useJackknife) {
+    val trainingLabels = trainingData.collect { case (_, y) if !(y.isInfinite || y.isNaN) => y }
     val zeroTolerance = StatsUtils.range(trainingLabels) / 1e12
     BaggerHelper.calculateRescaleRatio(oobErrors.map { case (_, e, u) => (e, u) }, zeroTolerance = zeroTolerance)
   } else {
@@ -63,9 +54,7 @@ protected case class BaggerHelper(
   }
   assert(!rescaleRatio.isNaN && !rescaleRatio.isInfinity, s"Uncertainty calibration ratio is not real: $rescaleRatio")
 
-  /**
-    * Data on which to train a bias learner.
-    */
+  /** Data on which to train a bias learner. */
   lazy val biasTraining: Seq[(Vector[Any], Double)] = oobErrors.map {
     case (f, e, u) =>
       // Math.E is only statistically correct.  It should be actualBags / Nib.transpose(i).count(_ == 0)
